[#17] Refactor code and use Factory for dependency injection
