+++ iosApp/PhongKMMIC/Sources/Application/KMMICApp.swift
@@ -11,7 +11,8 @@ import SwiftUI
 @main
 struct KMMICApp: App {
 
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    @UIApplicationDelegateAdaptor(AppDelegate.self)
+    var appDelegate
 
     var body: some Scene {
         WindowGroup {
