+++ iosApp/PhongKMMIC/Sources/Presentation/Coordinator/AppCoordinator.swift
@@ -25,37 +25,3 @@ struct AppCoordinator: View {
         .environmentObject(navigator)
     }
 }
-
-final class Navigator: ObservableObject {
-
-    @Published var routes: Routes<Screen> = [.root(.splash, embedInNavigationView: true)]
-
-    func showScreen(screen: Screen, with transition: Transition) {
-        switch transition {
-        case .root:
-            routes = [.root(screen, embedInNavigationView: true)]
-        case .push:
-            routes.push(screen)
-        case .presentSheet:
-            routes.presentSheet(screen)
-        case .presentCover:
-            routes.presentCover(screen, embedInNavigationView: true)
-        }
-    }
-
-    func goBack() {
-        routes.goBack()
-    }
-
-    func goBackToRoot() {
-        routes.goBackToRoot()
-    }
-
-    func pop() {
-        routes.pop()
-    }
-
-    func dismiss() {
-        routes.dismiss()
-    }
-}
